# server/openapi.yaml
openapi: 3.0.0
servers:
  - url: http://localhost:8080
    description: local test server
info:
  description: This is a simple API for miniGameStore
  version: "1.0.0"
  title: miniGameStore
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: ivan.maidaniuk@gmail.com
...
paths:
...
  /api/users:
    get:
      tags:
        - "User"
      summary: "Get users"
      operationId: "getUsers"
      parameters:
        - in: query
          name: offset
          allowEmptyValue: true
          schema:
            type: integer
          description: The number of items to skip before starting to collect the result set.
        - in: query
          name: limit
          allowEmptyValue: true
          schema:
            type: integer
          description: The number of items to return.
        - in: query
          name: search
          allowEmptyValue: true
          schema:
            type: string
          description: String for searching friends
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  offset:
                    type: integer
                  limit:
                    type: integer
                  search:
                    type: string
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/UnfamiliarUser"
        "4XX":
          description: "Client error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "5XX":
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - cookieAuth: []
        - bearerAuth: []
  /api/users/me:
    get:
      tags:
        - "User"
      summary: "Get user's user-profile info"
      description: "User can request only his/her own user-profile info"
      operationId: "getProfileInfo"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "4XX":
          description: "Client error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "5XX":
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - cookieAuth: []
        - bearerAuth: []
    delete:
      tags:
        - User
      summary: Delete user's user-profile
      description: User can delete only his/her own user-profile info
      operationId: "deleteUserProfile"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "4XX":
          description: "Client error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "5XX":
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - cookieAuth: []
        - bearerAuth: []
    patch:
      tags:
        - User
      summary: Edit user's user-profile
      description: User can Edit only his/her own user-profile info
      operationId: "editUserProfile"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              example:
                { birthday: "1999-10-28T08:03:19.814Z", username: "Guts_2007" }
              properties:
                username:
                  $ref: "#/components/schemas/username"
                birthday:
                  $ref: "#/components/schemas/date"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "4XX":
          description: "Client error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "5XX":
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - cookieAuth: []
        - bearerAuth: []
  /api/users/me/password:
    patch:
      tags:
        - User
      summary: Change user's password
      description: User can change only his/her own user-profile password
      operationId: "changeUserPassword"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example: { oldPassword: "123456", newPassword: "qwerty" }
              required:
                - oldPassword
                - newPassword
              properties:
                oldPassword:
                  $ref: "#/components/schemas/password"
                newPassword:
                  $ref: "#/components/schemas/password"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "4XX":
          description: "Client error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "5XX":
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - cookieAuth: []
        - bearerAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    _id:
      oneOf:
        - type: string
        - type: object
      example: "507f191e810c19729de860ea"
    password:
      type: string
      format: password
      minLength: 6
      maxLength: 20
      example: "qwerty"
...
    username:
      type: string
      nullable: true
      minLength: 2
      maxLength: 20
      example: "obiwan00"
    email:
      type: string
      format: email
      example: "obi@mail.com"
...
    date:
      type: string
      format: date-time
      example: "2000-10-28T08:03:19.814Z"
      description:
        'Use JavaScript method "toISOString" of Date prototype (Data.prototype.toISOString())
        to get date in right format'
    nullableDate:
      type: string
      nullable: true
      format: date-time
      example: "2000-10-28T08:03:19.814Z"
      description:
        'Use JavaScript method "toISOString" of Date prototype (Data.prototype.toISOString())
        to get date in right format'
...
    Success:
      type: object
      properties:
        message:
          type: string
          example: "Success message"
    User:
      required:
        - _id
        - email
        - createdAt
      properties:
        _id:
          $ref: "#/components/schemas/_id"
        username:
          $ref: "#/components/schemas/username"
        email:
          $ref: "#/components/schemas/email"
        birthday:
          $ref: "#/components/schemas/nullableDate"
        createdAt:
          $ref: "#/components/schemas/date"
...
    UnfamiliarUser:
      type: object
      required:
        - _id
      properties:
        _id:
          $ref: "#/components/schemas/_id"
        username:
          $ref: "#/components/schemas/username"
...
    Error:
      required:
        - message
      properties:
        message:
          type: string
          example: "Error message"